+++ src/chains/Arc/Common.sol
@@ -19,6 +19,7 @@ import {
     uniswapV3ForkId,
     IUniswapV3Callback
 } from "../../core/univ3forks/UniswapV3.sol";
+import {achSwapV3Factory, achSwapV3InitHash, achSwapV3ForkId} from "../../core/univ3forks/AchSwapV3.sol";
 import {ARC_POOL_MANAGER} from "../../core/UniswapV4Addresses.sol";
 
 // Solidity inheritance is stupid
@@ -67,6 +68,10 @@ abstract contract ArcMixin is FreeMemory, SettlerBase, UniswapV4 {
         if (forkId == uniswapV3ForkId) {
             factory = uniswapV3ArcFactory;
             initHash = uniswapV3InitHash;
+            callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
+        } else if (forkId == achSwapV3ForkId) {
+            factory = achSwapV3Factory;
+            initHash = achSwapV3InitHash;
             callbackSelector = uint32(IUniswapV3Callback.uniswapV3SwapCallback.selector);
         } else {
             revertUnknownForkId(forkId);